requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0

# Google Sheets integration
gspread>=5.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Authentication & Security
bcrypt>=4.0.1
cryptography>=41.0.0

# Data visualization
plotly>=5.17.0
altair>=5.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
